Avoid sending error message twice in BatchAction
BatchAction.handle() displays an aggregate error message
if an exception is returned from individual action() method.
If we don't need a detail error message, try/except is unnecessary.

Note that a fix for a case where a detail error message is needed
will be covered in the subsequent patches.

Change-Id: I8037355b12e0756c6add5c5b9ba05a430155dfe5
Partial-Bug: #1733207