Fixes the deletion on click action assignment
Changes the selector used to assign the delete on
click action to the correct button in the modal
confirmation dialog.

In addition, the code that hides the "working"
modal to the success handler of said delete action;
the modal should now appear while elements are
actually being deleted and disappearing after
the operation finishes.

Closes-Bug: 1733324
Change-Id: Ia14c263ed49c6f9705dbe12e8108a2c8c858aa31