Fix admin/project instance UT
The expected error count was wrongly changed in
commit df194c8b4c25c6128f03a98fd0e628ba14b743ff.

The expected error count was changed from 1 to 2 in the commit
but one new error message was caused by a wrong test code.
api.glance.image_list_detailed should return a tuple of
(images, has_more, has_prev) but the mocked method only returns
a list of images and this causes another error.

Change-Id: Ic1083014f9572bfd20adc60e7ec3eeeac387e9d5