Redirect VM console dynamically
Previously when a user accesses the network topology panel
horizon resolves VM console URL when returning server data.
This requires one API call to nova per VM, and this causes
a scaling issue.

Actually there is no need to resolves VM console URL when
the network topology. A console URL corresponding to a requested
VM can be resolved when a user actually accesses a console.
This reduces the number of API calls drastically and addresses
the scaling issue reported in the bug report.

Change-Id: Icc667449e2988e6227012a6e899835a97ce0d738
Closes-Bug: #1723142