Trunks panel: display the MAC of ports
Display some additional useful information while selecting ports for a
trunk. Since in a common use case the subports should use the same MAC
address as the parent port it is especially important to display the
MAC address. Displaying the device owner was requested in a previous
review.

Change-Id: Ic86dbded8f68659763c78b773146f2a61c5cb662
Partially-Implements: blueprint neutron-trunk-ui