Django2: Use new custom template loader interface
The new template loader interface is described here:
https://docs.djangoproject.com/en/2.0/ref/templates/api/#loader-methods

The old interface was deprecated in Django 1.9, and the old interface
load_template_sources() was removed in Django 2.0.
https://docs.djangoproject.com/en/2.0/releases/1.9/#template-loader-apis-have-changed

According to the UT result, force_escape in _detail_overview.html
looks unnecessary. If force_escape is used, the returned string will
be double-escaped and it is not what we expect.

blueprint django2-support
Change-Id: I8a5a79cb61a6ebd25453bde3e0ec64f5b5f50e1b