Drop flavor name validation in horizon side.
This patch drop name validation in horizon and
let nova check if the is name is valid or not.
Instead of using pattern in the flavors form,
pattern will be checked in the nova. Also I
added limation for the name size with tests to
check the max size since the min size has it's unit
test.

Closes-Bug: 1738328
Change-Id: Ib4f6f9b87560df43819d992254603e55ea8ff6a3