Trunks panel: simplify code for easier testing
The crossHide boolean flag was never a necessary part of the trunks
panel. The only downside of not having it is that we expose a data
structure between the step controllers even when that's not needed
(when crossHide previously was false).

Eliminating the if/else simplifies coverage testing in the followup
patch.

Change-Id: Id7b51997b48d2dd91057975c2fcfb2c40c33aa51
Partially-Implements: blueprint neutron-trunk-ui