font for transfer table heading enlarged
changed the heading type from h5 to h4 to enlarge it without
altering the appearance of other h5's. This was done because
the current font size made locating the text amongst the many
larger elements on the page, this change allows a user to
quickly identify the purpose of the sections below.

Closes-Bug #1743683

Change-Id: I992ed598d59e4fef6baa18139288782003a1ade3