Assign project role to the user when the user is created
In the angular create user form page, when the new user is created,
the role selected by the administrator is not assigned to the user,
which will result in the creation of the user can not log in normally.

This patch uses the existing service, when the new user is created,
then Assign project role for the new user.

Change-Id: I52b9bb3bc986cc89439cf22deb1e250a9252938a
Closes-Bug:#1743596