Handle novaclient ec2 cert support removal
In Queens, ec2 cert support was removed from python-novaclient.
This causes horizon to be unable to get the keystone values of
ec2 credentials.

This patch removes the code that handles ec2 certs from nova.
Since the ec2 cert code in nova was removed in Mitaka. There is
no real risk of breaking backwards compatibility.

Change-Id: I470761d67004f6e6b188d3afc4b7f081b7bc708f
Closes-Bug: #1729175