i18n: Allow translator to control the word order (trunk)
"<IP address> on subnet <subnet>" format in "Parent Port" and
"Subports" tabs in "Create Trunk" and "Edit Trunk" form did not
allow translators to control the word order because only "on subnet"
was marked as translation string and they are concatenated.

This commit improves the translation string by using the "translate
parameters" feature of Angular gettext.

Also "on subnet" in the "Network Ports" tab of "Create Instance"
workflow is now marked as translation string.

Change-Id: I189897eee8037466ace2b68b8f9be3dc242c4252
Closes-Bug: #1745873