i18n: Do not include html directives in translation strings
It is important not to include html directives in translation targets
as much as possible to avoid unnecessary errors when selecting
translated languages.

Note that this commit does not necessarily exclude all html directives
from translation strings. Some of them needs to be included to allow
translators to control the word order properly.
This commit mainly focuses on eliminate unrelated html directives.

Change-Id: I66a24d4b01e695aab226b7c1cd4bc4f50ab33a7f