Refactor the create port form to workflow
The neutron API allows the users to specify security group(s) when
creating a port [1], but at the moment the Horizon can only modify
the security group of a port by 'Edit Port' row action.

As with edit port workflow, refactoring the creation of port form
to workflow is to support the creation of ports as well as the
assignment of security groups. Since the change may be more, this
patch only contains reconstruct the form, support the security
groups will be submitted in another patch.

[1] https://developer.openstack.org/api-ref/network/v2/index.html#create-port

Change-Id: Ie7e5242cc885b6eb0316f695453b0d2844360d67
Related-bug: #1746985