Make create_mocks compatible with create_stubs
create_mocks was introduced to mock multiple methods easily,
but its arguments are not compatible with create_stubs for mox.
This requires us to make many many small changes when converting
mox tests into mock. It would be nice if create_mocks is compatible
with create_stubs.

Part of blueprint mock-framework-in-unit-tests
Change-Id: I3d98448b1b9c517c194d692d6b3b53dc00e46223