Add assert_mock_multiple_calls_with_same_arguments test helper
During the migration from mox to mock, it turns out there is case where
some method is called multiple times with same parameters and
it is common in openstack dashboard tests.

This commit adds assert_mock_multiple_calls_with_same_arguments assertion
method to ensure to check both call_count and arguments of all calls.

Note that assert_has_calls just checks a subset of called arguments
so it is better to check arguments of all calls.

Part of blueprint mock-framework-in-unit-tests
Change-Id: Id939e61718b0405a02a22a5ec8e55eda9554c545