Switch project instances tests to mock
Python3, map() returns a map object. It is better to ensure to pass
an interable object when the value is passed to another library.
It also breaks mock assert_has_calls/assert_called_once_with
because they fails to compare map object and an array.
This commit changes map object passed in
api.neutron.server_update_security_groups into an explicit array
to address this problem.

Change-Id: I4a45489cdbe2f5ca3b17e30f2e62159dd5ab74f8
Partially-Implements: blueprint mock-framework-in-unit-tests