Add same launch instance error handling as image to instance
Previously, when launching an instance with an instance snapshot rather
than with an image, the error handling was not as pleasant. I.e. the
volume size was not automatically set to the minimum size and the error
message would only appear after attempting to launch an instance. This
is a small patch that addresses this and makes the user experience more
friendly.

Change-Id: Ib694f66c74e90ee2e15201673de953c08cf10122
Closes-bug: #1752360