Support description for instance update/rebuild
In Nova Compute API microversion 2.19, you can specify a description
attribute when creating, rebuilding, or updating a server instance. This
description can be retrieved by getting server details, or list details
for servers, this patch adds support for this attribute for instance in
horizon.
This patch adds description for instance update/rebuild

Change-Id: I1c561607551fe6ed521772688b643cb27400e24e
Closes-bug: #1753661