Update upper-constraints consumption
In order to deal with horizon going into the upper-constraints file so
that horizon plugins can consume it normally, we need to align the
constraints invocation in horizon's tox.ini. This means putting the -C
argument into the dependencies list rather than the pip install line so
that it doesn't get appended to tox installing horizon itself.

While we're in here, move nosehtmloutput into test-requirements so that
we don't need to do .[test] which also doesn't work with constraints.

Needed-By: https://review.openstack.org/550475
Change-Id: I736f79010c2d34e25d41d967a38128c2263ed13f