The disk format is selected automatically when using file to create image
The disk format field has its onchange listener in horizon.forms.js file,
but it is valid only when the change is made by mouse or keyboard.
So, angularJs controller is just change its value and its listener is
not triggered. The listener can't be triggered manually within controller.
This commit set the display manually and resolve it.

Change-Id: I8c228bac9392003055a808eeb56b733ac4c9b07a
Closes-Bug: #1760006