doc: Update apache configuration for WSGI
It makes more sense to use <Location> directive rather than
<Directory> directive because HTTP requests are handled by
WSGI script and the directory paths are not actually referred.
The proposed approach is already used in DevStack and what we test.

This is a patch to follow up the discussion on the apache
configuration files in https://review.openstack.org/#/c/561802/.

Change-Id: I0c3111ed4d04c55ba31924d02bbf3eb416ec3412