Move to 404 page if specified navigation not found
When refresh or link directly to ngdetails without correct navigation
to be set, ngdetails view can not reproduce navigations, i.e. sidebar
and breadcrumb.
e.g. when the ordinary user try to open the URL specified as admin side.

In this situation, ngdetails view should show 404 page.
So this patch move to 404 page if the navigation does not exist.

Change-Id: I7aeeddf838923d110703bfd7a04699b056c2abd5
Closes-Bug: #1761036