Use new style for help in Angular ModalFormService
ModalFormService (horizon.framework.widgets.form.ModalFormService,
used in create key pair dialog or create user dialog) and WizardModalService
(horizon.framework.widgets.modal.wizard-modal.service, used in create instance
dialog or create image dialog) are used in dialogs for create/edit/etc in Angularized
panel. But their help descriptions are shown in different style.

Angularization proposes new look and feel, so ModalFormService should follow
new style for help description like WizardModalService.

This patch proposes new style for help in dialogs that uses ModalFormService.

And "?" button for opening help panel is overlaped on input field on the first row,
so this patch moves it up little bit.

In this new style, if the help contents is too simple, it is not needed to
make description space in the right harf of dialog for looks consistency,
it is prefer not to add help panel, i.e. simply do not show "?" button.

Change-Id: Ia720180c7aaef1f3d9c5e24dbebe3fe2d796d6c7
Closes-Bug: #1781151