Replaces the delete button with a disabling version
Replaces the default delete button template for the images batch delete
action with a custom one. This template contains a new component, which
changes the disabled/enabled state of the underlying action based on the
selected images allowed value as specified by the delete-image service
used to evaluate the allowed value for individual rows/images and their
respective delete actions.

Closes-Bug: 1549544

Change-Id: Idd5f59e32268cf4ceefe3639c607fb6c3520e538