add dns_name and dns_domain into admin column
https://review.openstack.org/#/c/561185/
According to the above merged commit, dns_name and
dns_domain are added into project/floating_ips/tables.py.

But because of the class 'FloatingIPsTable' inherits
the project_tables.FloatingIPsTable, so that we should
add dns_name and dns_domain into columns of Meta.

If not, if you click the Admin/Network/Floating IPs, the
page will crash.

Closes-Bug: #1785897
Change-Id: I6e933749a56a492989cab002c1b65656b7e2c65b