Add user tab in project details view.
Add an extra tab "users" in the project details view which
displays the users which have a role on the project (the
members of the project).

The users are displayed in a table which is an extension
(inheritance) of the user table used in the Users panel.

An extra column is added to this table, displaying the
roles of each user on project

Only users which have directly a role on the project
are displayed, the users which have a role through a
group are not displayed.

Change-Id: I88b40fcda300ee4640155347d479a972abb2df02
Partial-Bug: #1785263