Add groups tab in project details view.
Add an extra tab "Groups" in the project details view which
displays the groups which have a role on the project (the
groups members of the project).

The groups are displayed in a table which is an extension
(inheritance) of the group table used in the Groups panel.

An extra column is added to this table, displaying the
roles of each group on project

Change-Id: Ie6338c7e10b67900d5a553d0328b29bfd4a28212
Closes-Bug: #1785263