Revert "Network Topology graph "twitches""
This reverts commit e5dae9b35349c53fa6184e415e254094893cab1f
(and commit 1ca8555011ad345e6343eded2fcca4516aa70944 which is
a follow-up styling fix).

The reverted commit itself is a nice improvement,
but considering the situation that Rocky release is approaching
at the end of August we revert this to recover the original
user-visible behavior. It would be nice if someone can revisit
the improvement in the next Stein release.

Change-Id: If324bd47a1b8c1b9579c1f38adeae31fdf583780
Closes-Bug: #1786628
Related-Bug: #1716834