Drop mox usage in Selenium tests
Selenium tests now runs in our gate,
so this is the time to migrate them to mock.

The arguments of the first calls of image_list_detailed
in test_modal_create_image_from_url and
test_modal_create_image_from_file need to be updated.
It seems mox did not check arguments strictly.

blueprint mock-framework-in-unit-tests

Change-Id: If60734d6dbd667fe37ffe64858b93f503e6b8386