Drop use_mox from horizon test helpers
We announced in the Rocky cycle that use_mox will be dropped
in the beginning of Stein cycle. This commit drops use_mox and
cleans up mox related stuffs in horizon and openstack_dashboard codes.

Note that openstack_auth is under the mock migration,
so we cannot mox3 from test-requirements.txt.

blueprint mock-framework-in-unit-tests

Change-Id: I189daf2d678bef0fa8219acaab5cf45bbe54178f