Fix bug of updating metadata for images
Glance project has a blueprint hidden-images. The commit can be
seen from https://review.openstack.org/#/c/578755/
The os_hidden is added into image metadata as bool type. From
horizon, the string value of os_hidden is passed to glanceclient.
https://github.com/openstack/python-glanceclient/blob/master/glanceclient/v2/images.py#L341
But type is not correspond so raise exception to horizon from
glanceclient.

Here to change the type of os_hidden before passing it to glanceclient.

Closes-Bug: #1789172
Change-Id: I0d4edaa4b661d2da79cdb45e4afa97d826aeb65d