Avoid UnhashableKeyWarning in api.nova.novaclient
In python3, novaclient APIVersion instance is not hashable.
If APIVersion insstance is passed to novaclient(),
UnhashableKeyWarning will be emitted.
An error message on UnhashableKeyWarning is emitted repeatedly
and this can pollute error log with unuseful messages.

To convert all unhashable arguments into hashable variables,
a new decorator memoized_with_argconv is introduced.
This decorator takes a converter function.

Change-Id: I773355b9332b3b195576b51cc81eda80aa4402ed
Closes-Bug: #1790929