Select correct boot source when launching instance from different sources
When launching instance from snapshot, volume or volume snapshot,
the selected boot source was always image.
This was due to a reset to image boot source when receiving "change allowed boot
sources" event.
From now:
 - The "change allowed boot source" event does not flush a preselection
if the preselection boot source type is in the allowed sources.
 - Set the bootsource with preselection update correctly the form for each
different type of boot source.

Change-Id: I6907652731fa89b303d997fe2e3c331a68f085b6
Closes-bug: #1608565