Fix changing user's own password
The original code is monkey-patching keystoneclient object to add a
user_id attribute to it. This no longer works with more recent
versions of keystoneclient, as they wrap the client in a helper
class.

I'm not happy with this solution, it's likely to have side effects
and to break again. I'm putting it up for discussion for a better
solution.

Change-Id: Idb296d1b10fa02a0b4852e96fe8cb2bdd70380e0
Closes-bug: #1776678