Only thorw one prompt when resizing instance error
The horizon.workflow.Workflow will give a prompt when
method handle return False. And the exceptions.handle
also do it and the prompt is more detailed.
It is reasonable that one error throw one prompt.
Now, we only throw a detailed prompt by workflow.

The all service will return an reason contained in response
when the service could not handle correctly. And the reason
will be wrapped into a ClientException.
Resizing reasons are somethings, like:
 - No valid host was found. No valid host found for resize.
 - No valid host was found. There are not enough hosts available.
 ...

Change-Id: Id48eb236c293ff7296952f6371ac9e9cdb15f0c7
Closes-Bug: #1800579