Server overview: display hypervisor name if available
Previously this had a check for is_superuser.  That's sort of
right since the default nova-api policy doesn't return the hypervisor
if you don't have the admin priv.

But, since there is a modifiable policy enforced by the api,
let's just let that decide.  If nova-api returns the hypervisor,
presumably it has decided that we're allowed to see it, so let's
see it!

Change-Id: I0cfe9090e8263f983fa5f42f42616a26407be47a