Clarify middleware public methods
In Django 1.10 the new middleware interface was introduced [1].
process_request and process_response are no longer part of the interface.
On the other hand, process_exception is the part of the interface.

To clarify that we are using the new middleware interface,
it looks better to rename the methods in the old interface to private.

[1] https://docs.djangoproject.com/en/2.0/topics/http/middleware/

Change-Id: I0bb39321f3659458a86beb2b258e9e903392620a