Document Django and Python support policy
As discussed in the Stein PTG at Denver,
we support maintained LTS versions of Django and other maintained
Django versions are optional. For example, upcoming LTS version of
Django 2.2 will be supported. Django 2.1 support will be best-effort.

The policy for Python supported version is documented.
Python 3.6 is now covered by unit tests, so it can be considered
as experimental. Python 3.7 support will be added along with
the community goal.

Change-Id: I1d844b0333b2701adbae7a51033091bff1458313